requirements.txt
# Africa-USA Trade Intelligence Platform
# Requirements for 100% FREE operation - NO PAID APIs

# Core MCP and Web Framework
mcp>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
streamlit>=1.28.0

# Data Processing and Analysis
pandas>=2.1.0
numpy>=1.24.0
requests>=2.31.0
httpx>=0.25.0
aiohttp>=3.9.0

# Database and ORM
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.0  # PostgreSQL driver

# Web Scraping and Data Collection (FREE)
beautifulsoup4>=4.12.0  # HTML parsing
selenium>=4.15.0  # Advanced web automation
feedparser>=6.0.10  # RSS feed parsing
lxml>=4.9.0  # XML parsing

# Social Media APIs (FREE TIERS ONLY)
# python-linkedin>=0.9.4  # LinkedIn API (if needed)
# tweepy>=4.14.0  # Twitter API (if needed)
# instagrapi>=1.19.0  # Instagram API (if needed)

# Content Creation and NLP (FREE)
transformers>=4.35.0  # Free NLP models via Hugging Face

# Data Visualization
plotly>=5.17.0
matplotlib>=3.8.0
seaborn>=0.13.0

# Task Scheduling and Automation
apscheduler>=3.10.0
celery>=5.3.0
redis>=5.0.0

# API and Web Security
python-jose>=3.3.0
passlib>=1.7.0
python-multipart>=0.0.6

# Environment and Configuration
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Monitoring and Logging
loguru>=0.7.0
prometheus-client>=0.19.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Data Export and Reporting
openpyxl>=3.1.0
reportlab>=4.0.0
jinja2>=3.1.0

# FREE Notifications ONLY
python-telegram-bot>=20.7.0  # Telegram (100% free)
# NO SendGrid - use SMTP instead
# NO Twilio - use Telegram instead
# NO OpenAI - use Hugging Face instead

# Email via SMTP (FREE)
email-validator>=2.1.0

# Additional free data sources
yfinance>=0.2.22  # Free Yahoo Finance data

# JSON processing for API responses
json5>=0.9.14

# OAuth support for social media APIs
requests-oauthlib>=1.3.1

# NO PAID APIS:
# ❌ openai (costs money after $5 credit)
# ❌ sendgrid (limited free tier)
# ❌ twilio (only trial credits)
# ❌ newsapi (very limited free tier)
# ❌ openweathermap (requires API key)
# ❌ exchangerate-api (limited requests)